moving isRepresentable into the matrix generators
